requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.3
cycler==0.10.0
cython==0.28.4
Django==2.0.7
dtaidistance==1.1.2
idna==2.7
kiwisolver==1.0.1
numpy==1.15.0
pandas==0.23.3
pycparser==2.18
PyMySQL==0.9.2
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
python-gettext==3.0
mysqlclient==1.3.13
scikit-learn==0.19.1
scipy==1.1.0